Made Authenticated Layout Reusable
