+++ resources/js/types/payroll.d.ts
@@ -1,0 +1,10 @@
+interface Item {
+    label: string;
+    url: string;
+    icon: React.ComponentType<LucideProps>;
+}
+
+export interface navigationLinks {
+    title: string;
+    items: Item[];
+}
